20260421-fasemcp-googleauth-bq/pyproject.toml
[project]
name = "20260419-fastmcp-googleauth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.4",
    "google-cloud-bigquery>=3.41.0",
]
